反应堆案例测试

+++ 04多路io复用-poll-epoll-epoll反应堆/06案例epoll反应堆/wrap.cpp
@@ -57,7 +57,7 @@ int Socket(int family, int type, int protocol)
 	return ret;
 }
 
-int Read(int fd, void * ptr, size_t nbytes)
+int Read(int fd, void * ptr, unsigned int nbytes)
 {
 	int ret;
 again:
